user-service/requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
python-jose[cryptography]